refactor: 유틸 하위 폴더를 processor 패키지 하위로 이동
