+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/processor/utils/weatherRiskFilter/WeatherRiskFilter.java
@@ -1,4 +1,4 @@
-package com.ImSnacks.NyeoreumnagiBatch.utils.weatherRiskFilter;
+package com.ImSnacks.NyeoreumnagiBatch.processor.utils.weatherRiskFilter;
 
 import com.ImSnacks.NyeoreumnagiBatch.reader.dto.VilageFcstResponseDto;
 import com.ImSnacks.NyeoreumnagiBatch.writer.dto.ShortTermWeatherDto;
